Implemented the chess board detector and commented out unimplemented features
