+++ sensor/intrinsic_camera_calibrator/intrinsic_camera_calibrator/intrinsic_camera_calibrator/board_detectors/board_detector_factory.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 
 
-from intrinsic_camera_calibrator.board_detectors.apriltag_grid_detector import ApriltagGridDetector
+# from intrinsic_camera_calibrator.board_detectors.apriltag_grid_detector import ApriltagGridDetector TODO(knzo25): implement this later
 from intrinsic_camera_calibrator.board_detectors.board_detector import BoardDetector
 from intrinsic_camera_calibrator.board_detectors.chessboard_detector import ChessBoardDetector
 from intrinsic_camera_calibrator.board_detectors.dotboard_detector import DotBoardDetector
@@ -27,6 +27,6 @@ def make_detector(board_type: BoardEnum, **kwargs) -> BoardDetector:
     classes_dic = {
         BoardEnum.CHESSBOARD: ChessBoardDetector,
         BoardEnum.DOTBOARD: DotBoardDetector,
-        BoardEnum.APRILTAGGRID: ApriltagGridDetector,
+        # BoardEnum.APRILTAGGRID: ApriltagGridDetector, TODO(knzo25): implement this later
     }
     return classes_dic[board_type](**kwargs)
